fix: added the possibility to load custom plugins from a local folder
