+++ packages/cli/test/custom/monokle.validation.yaml
@@ -1,0 +1,4 @@
+plugins:
+  argo: true
+rules:
+  argo/argo-config-maps: "warn"
