+++ packages/validation/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@monokle/validation",
-  "version": "0.10.3",
+  "version": "0.11.0",
   "description": "Kubernetes resource validation",
   "author": "Kubeshop",
   "license": "MIT",
