+++ packages/validation/src/MonokleValidator.ts
@@ -8,19 +8,19 @@ import type {
   CustomSchema,
   Incremental,
   Plugin,
-  Resource,
+  Resource
 } from "./common/types.js";
 import { Config, PluginMap } from "./config/parse.js";
 import {
   PluginMetadataWithConfig,
   PluginName,
   RuleMetadataWithConfig,
-  Validator,
+  Validator
 } from "./types.js";
 import { nextTick, throwIfAborted } from "./utils/abort.js";
 import {
   extractSchema,
-  findDefaultVersion,
+  findDefaultVersion
 } from "./utils/customResourceDefinitions.js";
 import { PluginLoadError } from "./utils/error.js";
 import { isDefined } from "./utils/isDefined.js";
@@ -33,6 +33,8 @@ import { RemoteWasmLoader } from "./validators/open-policy-agent/index.js";
 import { OpenPolicyAgentValidator } from "./validators/open-policy-agent/validator.js";
 import { ResourceLinksValidator } from "./validators/resource-links/validator.js";
 import { YamlValidator } from "./validators/yaml-syntax/validator.js";
+import * as fs from "fs";
+import path from "path";
 
 export type PluginLoader = (name: string) => Promise<Plugin>;
 
@@ -70,9 +72,7 @@ export function createExtensibleMonokleValidator(
         return new DevCustomValidator(parser);
       default:
         try {
-          const url = `https://plugins.monokle.com/validation/${pluginName}/latest.js`;
-          const customPlugin = await import(/* @vite-ignore */ url);
-          return new SimpleCustomValidator(customPlugin.default, parser);
+          return await loadExternalPlugin( pluginName, parser );
         } catch (err) {
           throw new Error(
             err instanceof Error
@@ -115,6 +115,33 @@ export function createDefaultPluginLoader(
   };
 }
 
+export async function loadExternalPlugin(pluginName: string, parser: ResourceParser) {
+  const filePath = path.join(process.cwd(), ".monokle-plugins", `${pluginName}-plugin.js`);
+  if (fs.existsSync(filePath)) {
+    const customPlugin = await import(/* @vite-ignore */ filePath );
+    return new SimpleCustomValidator(customPlugin.default, parser);
+  } else {
+    const url = `https://plugins.monokle.com/validation/${pluginName}/latest.js`;
+    const customPlugin = await importWithDataUrl(url);
+    return new SimpleCustomValidator(customPlugin.default, parser);
+  }
+}
+
+
+async function importWithDataUrl(url: string) {
+  const response = await fetch(url);
+  if (!response.ok) {
+    throw new Error(`Error fetching ${url}: ${response.statusText}`);
+  }
+  const source = await response.text();
+  const buff = Buffer.from(source);
+  const encodedSource = buff.toString("base64");
+  const dataUrl = `data:text/javascript;base64,${encodedSource}`;
+
+  return await import(/* @vite-ignore */ dataUrl);
+}
+
+
 /**
  * The plugins that will be loaded by default.
  */
@@ -122,7 +149,7 @@ const DEFAULT_PLUGIN_MAP = {
   "open-policy-agent": true,
   "resource-links": true,
   "yaml-syntax": true,
-  "kubernetes-schema": true,
+  "kubernetes-schema": true
 };
 
 export class MonokleValidator implements Validator {
@@ -310,7 +337,7 @@ export class MonokleValidator implements Validator {
       this.#plugins.map((p) =>
         p.configure({
           rules: config.rules,
-          settings: config.settings,
+          settings: config.settings
         })
       )
     );
@@ -331,10 +358,10 @@ export class MonokleValidator implements Validator {
    * Validates the resources.
    */
   async validate({
-    resources,
-    incremental,
-    abortSignal: externalAbortSignal,
-  }: {
+                   resources,
+                   incremental,
+                   abortSignal: externalAbortSignal
+                 }: {
     resources: Resource[];
     incremental?: Incremental;
     abortSignal?: AbortSignal;
@@ -370,7 +397,7 @@ export class MonokleValidator implements Validator {
     return {
       $schema: "https://json.schemastore.org/sarif-2.1.0.json",
       version: "2.1.0",
-      runs,
+      runs
     };
   }
 
@@ -399,7 +426,7 @@ export class MonokleValidator implements Validator {
   async registerCustomSchema(schema: CustomSchema) {
     if (!this.isPluginLoaded("kubernetes-schema")) {
       this.debug("Cannot register custom schema.", {
-        reason: "Kubernetes Schema plugin must be loaded.",
+        reason: "Kubernetes Schema plugin must be loaded."
       });
       return;
     }
@@ -407,7 +434,7 @@ export class MonokleValidator implements Validator {
     const key = `${schema.apiVersion}-${schema.kind}`;
     if (this.#customSchemas.has(key)) {
       this.debug("Cannot register custom schema.", {
-        reason: "The schema is already registered.",
+        reason: "The schema is already registered."
       });
       return;
     }
@@ -428,7 +455,7 @@ export class MonokleValidator implements Validator {
   async unregisterCustomSchema(schema: Omit<CustomSchema, "schema">) {
     if (!this.isPluginLoaded("kubernetes-schema")) {
       this.debug("Cannot unregister custom schema.", {
-        reason: "Kubernetes Schema plugin must be loaded.",
+        reason: "Kubernetes Schema plugin must be loaded."
       });
       return;
     }
@@ -436,7 +463,7 @@ export class MonokleValidator implements Validator {
     const key = `${schema.apiVersion}-${schema.kind}`;
     if (this.#customSchemas.has(key)) {
       this.debug("Cannot register custom schema.", {
-        reason: "The schema is not registered.",
+        reason: "The schema is not registered."
       });
       return;
     }
